requirements.txt
###### Requirements without Version Specifiers ######`
pyzbar
pillow
###### Requirements with Version Specifiers ######`

